Fix a critical misunderstanding of how writing works.
The code was misread and we believe that partitions are all read from a single input stream. In fact this is completely contrary to the way it works; we're merging the contents of spill files together that have the same matching partition ID. The API is updated to reflect this.